feat(ci): upgrade .NET SDK to version 9.0
Updated the .gitlab-ci.yml to use the .NET SDK version 9.0 for both build and deploy stages. Modified the Panda.NuGet.BillbeeClient.csproj to target .NET 9.0 alongside existing frameworks. Additionally, added a package reference for Microsoft.Extensions.Http for the .NET 9.0 target framework to ensure compatibility with the latest SDK features. This upgrade prepares the project for future development with the latest tools and dependencies.